+++ zetryn/__init__.py
@@ -19,7 +19,7 @@ from .core import (
     StepTrace,
 )
 
-__version__ = "0.9.0"
+__version__ = "0.10.0"
 
 __all__ = [
     "END",